use evemu; DROP TABLE IF EXISTS srvInvTypesExtra; CREATE TABLE srvInvTypesExtra (     typeID INT,     soundID INT,     dataID INT,     typeNameID INT,     descriptionID INT,     PRIMARY KEY (typeID),     INDEX soundID (soundID),     INDEX dataID (dataID),     INDEX typeNameID (typeNameID),     INDEX descriptionID (descriptionID) ); INSERT INTO srvInvTypesExtra (typeID, soundID, dataID, typeNameID, descriptionID) SELECT invTypes.typeID, invTypes.soundID, invTypes.dataID, invTypes.typeNameID, invTypes.descriptionID FROM invTypes; ALTER TABLE `invTypes` DROP COLUMN `descriptionID`, DROP COLUMN `typeNameID`, DROP COLUMN `dataID`, DROP COLUMN `soundID`, CHANGE COLUMN `basePrice` `basePrice` DOUBLE NULL DEFAULT NULL , DROP INDEX `descriptionID` , DROP INDEX `typeNameID` , DROP INDEX `dataID` , DROP INDEX `soundID` ; update invTypes set basePrice = basePrice / 10000.0 where typeID>0;

+++ src/eve-server/corporation/CorpStationMgrService.cpp
@@ -226,7 +226,7 @@ PyResult CorpStationMgrIMBound::Handle_SetCloneTypeID(PyCallArgs &call) {
     }
 
     //Get cost of clone
-    int cost = CorporationDB::GetCloneTypeCostByID(arg.CloneTypeID);
+    double cost = CorporationDB::GetCloneTypeCostByID(arg.CloneTypeID);
 
     //Check if player has enough money
     if(call.client->GetBalance() > cost) {
